+++ src/OmnichannelChatSDK.ts
@@ -1295,12 +1295,17 @@ class OmnichannelChatSDK {
                     ChatId: this.chatToken.chatId as string
                 });
             } catch (error) {
+                const exceptionDetails: ChatSDKExceptionDetails = {
+                    response: ChatSDKErrorName.ChatSDKSendMessageFailed,
+                    errorObject: `${error}`
+                };
                 this.scenarioMarker.failScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
                 });
 
-                throw new Error('ChatSDKSendMessageFailed');
+                throw new Error(ChatSDKErrorName.ChatSDKSendMessageFailed);
             }
         } else {
             const messageToSend: IRawMessage = {
@@ -1333,13 +1338,18 @@ class OmnichannelChatSDK {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string
                 });
-            } catch {
+            } catch (error) {
+                const exceptionDetails: ChatSDKExceptionDetails = {
+                    response: ChatSDKErrorName.ChatSDKSendMessageFailed,
+                    errorObject: `${error}`
+                };
                 this.scenarioMarker.failScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
-                    ChatId: this.chatToken.chatId as string
+                    ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
                 });
 
-                throw new Error('ChatSDKSendMessageFailed');
+                throw new Error(ChatSDKErrorName.ChatSDKSendMessageFailed);
             }
         }
     }